backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==21.2.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.0
httpx==0.27.0
pynacl==1.5.0
bech32==1.2.0

